terminal modifications
