+++ src/App.css
@@ -70,3 +70,9 @@ code {
   height: 16px;
   margin-right: 10px;
 }
+
+#terminal {
+  height: 500px;
+  overflow-y: scroll;
+  overflow-x: hidden;
+}
